LayoutTests/fast/dom/Element/attrisid-extra01.html

<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>

<script>
description("Test Attr.isId where Attr is created without element.");
    
var attr = document.createAttributeNS("http://www.w3.org/XML/1998/namespace","xml:lang");
shouldBeFalse('attr.isId');

var attr = document.createAttributeNS("","id");
shouldBeTrue('attr.isId');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
