safari-534.53-branch/LayoutTests/fast/dom/Element/attrisid-extra01-expected.txt

Test Attr.isId where Attr is created without element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS attr.isId is false
PASS attr.isId is true
PASS successfullyParsed is true

TEST COMPLETE
